refactor: move version prompt enable/disable to a config subcommand, similar to git

+++ src/algokit/core/version_prompt.py
@@ -16,18 +16,12 @@ LATEST_URL = "https://api.github.com/repos/algorandfoundation/algokit-cli/releas
 VERSION_CHECK_INTERVAL = timedelta(weeks=1).total_seconds()
 
 
-@click.pass_context
-def do_version_check(context: click.Context, _: None) -> None:
-    _do_version_check(context)
-
-
-# this second function is to allow mocking
-def _do_version_check(context: click.Context) -> None:
-    if context.obj["skip_version_check"] or _is_version_check_disabled():
-        logger.debug("Skipping version check")
+def do_version_prompt() -> None:
+    if _is_version_prompt_disabled():
+        logger.debug("Version prompt disabled")
         return
 
-    current_version = _parse_version(metadata.version(PACKAGE_NAME))
+    current_version = get_current_version()
     if current_version is None:
         logger.debug("Could not determine current version")
         return
@@ -39,14 +33,14 @@ def _do_version_check(context: click.Context) -> None:
 
     if current_version < latest_version:
         logger.info(
-            f"You are using AlgoKit version {_version_str(current_version)}, "
-            f"however version {_version_str(latest_version)} is available."
+            f"You are using AlgoKit version {format_version(current_version)}, "
+            f"however version {format_version(latest_version)} is available."
         )
     else:
         logger.debug("Current version is up to date")
 
 
-def _version_str(version: Version) -> str:
+def format_version(version: Version) -> str:
     return "" if version is None else ".".join(map(str, version))
 
 
@@ -76,17 +70,21 @@ def get_latest_version_or_cached() -> Version:
     if last_checked is None or (now - last_checked) > VERSION_CHECK_INTERVAL:
         version = get_latest_version()
         last_checked = now
-        version_check_path.write_text(_version_str(version), encoding="utf-8")
+        version_check_path.write_text(format_version(version), encoding="utf-8")
     return version
+
+
+def get_current_version() -> Version:
+    return _parse_version(metadata.version(PACKAGE_NAME))
 
 
 def get_latest_version() -> Version:
     headers = {"ACCEPT": "application/vnd.github+json", "X-GitHub-Api-Version": "2022-11-28"}
-    # while algokit is private we need a token to query the release
+    # TODO: remove GH_TOKEN auth once algokit repo is public
     gh_token = os.getenv("GH_TOKEN")
     if gh_token:
-        logger.debug("Using GH_TOKEN to query latest version")
         headers["Authorization"] = f"Bearer {gh_token}"
+
     response = httpx.get(LATEST_URL, headers=headers)
     if response.status_code != 200:
         return None
@@ -100,47 +98,30 @@ def get_latest_version() -> Version:
     return tuple(map(int, match.group(1).split(".")))  # type: ignore[return-value]
 
 
-def _set_version_check(_ctx: click.Context, _: click.Option, value: bool | None) -> None:
-    if value is not None:
-        _store_version_check(enable=value)
-        _ctx.exit()
-
-
-def _skip_version_check(_ctx: click.Context, _: click.Option, value: bool) -> None:  # noqa FBT001
-    _ctx.ensure_object(dict)
-    _ctx.obj["skip_version_check"] = value
-
-
-def _is_version_check_disabled() -> bool:
-    disable_marker = get_app_config_dir() / "disable-version-check"
+def _is_version_prompt_disabled() -> bool:
+    disable_marker = get_app_config_dir() / "disable-version-prompt"
     return disable_marker.exists()
 
-
-def _store_version_check(*, enable: bool) -> None:
-    disable_marker = get_app_config_dir() / "disable-version-check"
-    if enable:
-        disable_marker.unlink(missing_ok=True)
-        logger.info("Version check enabled")
-    else:
-        disable_marker.touch()
-        logger.info("Version check disabled")
-
-
-version_check_option = click.option(
-    "--enable-version-check/--disable-version-check",
-    callback=_set_version_check,
-    show_default=False,
-    default=None,
-    expose_value=False,
-    help="Enable or disable version checking and reporting permanently.",
-)
 
 skip_version_check_option = click.option(
     "--skip-version-check",
     is_flag=True,
-    callback=_skip_version_check,
     show_default=False,
     default=False,
-    expose_value=False,
-    help="Skip version checking and reporting.",
+    help="Skip version checking and prompting.",
 )
+
+
+@click.command("version-prompt", short_help="Enables or disables version prompt")
+@click.argument("enable", required=False, type=bool, default=None)
+def version_prompt_configuration_command(*, enable: bool | None) -> None:
+    if enable is not None:
+        disable_marker = get_app_config_dir() / "disable-version-prompt"
+        if enable:
+            disable_marker.unlink(missing_ok=True)
+            logger.info("📡 Resuming check for new versions")
+        else:
+            disable_marker.touch()
+            logger.info("🚫 Will stop checking for new versions")
+    else:
+        logger.info(str(not _is_version_prompt_disabled()))
